much needed separation of backend and file processor
